Librairie pour le classification et la regression personnalisées/codebase/requirements.txt
absl-py==0.12.0
-e git+https://github.com/criteo-research/pytorch-ada.git@ac924cc0780d0d914982950d05b19b4975752c9d#egg=ada&subdirectory=adalib
alembic==1.4.1
antlr4-python3-runtime==4.8
anyio @ file:///D:/bld/anyio_1614388928981/work/dist
appdirs==1.4.4
argon2-cffi @ file:///D:/bld/argon2-cffi_1605217273648/work
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
Babel @ file:///home/conda/feedstock_root/build_artifacts/babel_1605182336601/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1616843134209/work
black==20.8b1
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1612213472466/work
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi @ file:///D:/bld/cffi_1613414140225/work
chardet @ file:///D:/bld/chardet_1610093646853/work
click==7.1.2
cloudpickle==1.6.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cryptography @ file:///D:/bld/cryptography_1616851649655/work
cycler==0.10.0
Cython==0.29.22
databricks-cli==0.14.3
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1617520196422/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
docker==4.4.4
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
et-xmlfile==1.0.1
Flask==1.1.2
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.4
graphviz==0.16
grpcio==1.36.1
hydra-core==1.0.6
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
imbalanced-learn==0.8.0
importlib-metadata @ file:///D:/bld/importlib-metadata_1617116938855/work
importlib-resources==5.1.2
ipykernel @ file:///D:/bld/ipykernel_1617137223378/work/dist/ipykernel-5.5.3-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1616837024736/work
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jdcal==1.4.1
jedi @ file:///D:/bld/jedi_1610146948662/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1612119311452/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1600692310011/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1614815863336/work
jupyter==1.0.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1615693636836/work
jupyter-console==6.4.0
jupyter-core @ file:///D:/bld/jupyter_core_1612125450516/work
jupyter-packaging @ file:///home/conda/feedstock_root/build_artifacts/jupyter-packaging_1613054948399/work
jupyter-server==1.5.1
jupyterlab @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_1616195786210/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1617134349851/work
jupyterlab-widgets==1.0.0
kiwisolver @ file:///C:/ci/kiwisolver_1612282606037/work
llvmlite==0.36.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe @ file:///D:/bld/markupsafe_1610127683324/work
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mistune @ file:///D:/bld/mistune_1610113104461/work
mlflow==1.15.0
mypy-extensions==0.4.3
nbclassic @ file:///home/conda/feedstock_root/build_artifacts/nbclassic_1610352513187/work
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1614336084111/work
nbconvert @ file:///D:/bld/nbconvert_1605402041690/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
notebook @ file:///D:/bld/notebook_1616419472987/work
numba==0.53.1
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
olefile==0.46
omegaconf==2.0.6
openpyxl @ file:///tmp/build/80754af9/openpyxl_1615411699337/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandas @ file:///C:/ci/pandas_1614711443373/work
pandocfilters==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pathspec==0.8.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///C:/ci/pillow_1617386341487/work
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1617387394212/work
prometheus-flask-exporter==0.18.1
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1616432837031/work
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1615243893546/work
pynndescent==0.5.2
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///D:/bld/pyrsistent_1610147004705/work
PySocks @ file:///D:/bld/pysocks_1610291742246/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-editor==1.0.4
pytorch-lightning==0.7.3
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
#pywin32==227
pywinpty @ file:///D:/bld/pywinpty_1602377663724/work
PyYAML==5.4.1
pyzmq @ file:///D:/bld/pyzmq_1614611907285/work
qgrid==1.3.1
qtconsole==5.0.3
QtPy==1.9.0
querystring-parser==1.2.4
regex==2021.3.17
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn @ file:///C:/ci/scikit-learn_1614446896245/work
scipy @ file:///C:/ci/scipy_1616703433439/work
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
Send2Trash==1.5.0
#sip==4.19.13
six @ file:///C:/ci/six_1605187374963/work
smmap==4.0.0
sniffio @ file:///D:/bld/sniffio_1610318539925/work
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado @ file:///D:/bld/terminado_1617048799047/work
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.2
torchviz==0.0.2
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.59.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
typed-ast==1.4.2
typing-extensions==3.7.4.3
umap-learn==0.5.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1615828766818/work
waitress==2.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton @ file:///D:/bld/win_inet_pton_1610147566231/work
wincertstore==0.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
